Match func_002cb6c0 shop menu list and input from C
Rewrite over a ShopMenuWork overlay with real callee contracts
(func_002e2740/func_002e26a0 return int, func_002b2d50 takes s16,
func_002e0970 takes FclDrawColor), declare the pad words and the
D_0063FA40 spin entry as arrays so they are addressed absolutely, and
share the ShopSpinParam view with func_002db400.

+++ src/Yajima/y_list.c
@@ -123,7 +123,7 @@ void func_002e2700(void *arg0) {
 }
 
 // FUN_002E2740
-s16 func_002e2740(s32 arg0) {
+s32 func_002e2740(s32 arg0) {
     u8 *g = iGpffffb58c;
     u8 *p;
     s32 count;
